+++ frontends/web/src/api/account.ts
@@ -348,6 +348,7 @@ export const proposeTx = (
 
 export type TSendTx = {
   success: true;
+  txId: string;
 } | {
   success: false;
   aborted: true;