+++ frontends/web/src/routes/router.tsx
@@ -164,31 +164,40 @@ export const AppRouter = ({ devices, deviceIDs, devicesKey, accounts, activeAcco
       accounts={activeAccounts} />
   </InjectParams>);
 
-  const BTCDirectEl = (<InjectParams>
+  const BTCDirectBuyEl = (<InjectParams>
     <BTCDirect
-      code={''}
-      accounts={activeAccounts} />
+      accounts={activeAccounts}
+      action="buy"
+      code={''} />
+  </InjectParams>);
+
+  const BTCDirectSellEl = (<InjectParams>
+    <BTCDirect
+      accounts={activeAccounts}
+      action="sell"
+      code={''} />
   </InjectParams>);
 
   const ExchangeEl = (<InjectParams>
     <Exchange
-      code={''}
       accounts={activeAccounts}
-      deviceIDs={deviceIDs}
+      code={''}
     />
   </InjectParams>);
 
   const PocketBuyEl = (<InjectParams>
     <Pocket
-      code={''}
       action="buy"
+      code={''}
+      deviceIDs={deviceIDs}
     />
   </InjectParams>);
 
   const PocketSellEl = (<InjectParams>
     <Pocket
-      code={''}
       action="sell"
+      code={''}
+      deviceIDs={deviceIDs}
     />
   </InjectParams>);
 
@@ -257,7 +266,8 @@ export const AppRouter = ({ devices, deviceIDs, devicesKey, accounts, activeAcco
             <Route index element={ExchangeInfoEl} />
             <Route path=":code" element={ExchangeInfoEl} />
           </Route>
-          <Route path="btcdirect/buy/:code" element={BTCDirectEl} />
+          <Route path="btcdirect/buy/:code" element={BTCDirectBuyEl} />
+          <Route path="btcdirect/sell/:code" element={BTCDirectSellEl} />
           <Route path="moonpay/buy/:code" element={MoonpayEl} />
           <Route path="pocket/buy/:code" element={PocketBuyEl} />
           <Route path="pocket/sell/:code" element={PocketSellEl} />