+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - iOS: Fix blank screens after prolonged inactivity
+- Added BTC Direct sell option
 
 ## v4.48.1
 - Bundle BitBox02 firmware version v9.23.1